backend/requirements.txt
fastapi
uvicorn
sqlalchemy
PyMuPDF
requests
openai==0.28
python-dotenv